convert linkUrlFilter to a Twig extension class
